remove unused structs move resources to .colony dir
