+++ internal/constants/constants.go
@@ -6,6 +6,7 @@ const (
 	KubeconfigHostPath     = "kubeconfig"
 	KubeconfigDockerPath   = "/output/kubeconfig"
 	ColonyYamlPath         = "colony.yaml"
+	ColonyDir              = ".colony"
 	ColonyNamespace        = "tink-system"
 	ColonyAPISecretName    = "colony-api"
 )
