+++ internal/docker/docker.go
@@ -4,10 +4,8 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"html/template"
 	"io"
 	"os"
-	"path/filepath"
 	"strings"
 	"time"
 
@@ -17,7 +15,6 @@ import (
 	"github.com/docker/docker/api/types/mount"
 	"github.com/docker/docker/client"
 	"github.com/konstructio/colony/internal/constants"
-	"github.com/konstructio/colony/internal/download"
 	"github.com/konstructio/colony/internal/logger"
 )
 
@@ -26,11 +23,6 @@ var ErrK3sContainerNotFound = errors.New("colony k3s container not found")
 type Client struct {
 	cli *client.Client
 	log *logger.Logger
-}
-
-type ColonyTokens struct {
-	LoadBalancerIP        string
-	LoadBalancerInterface string
 }
 
 func New(logger *logger.Logger) (*Client, error) {
@@ -96,27 +88,8 @@ func (c *Client) RemoveColonyK3sContainer(ctx context.Context) error {
 	return nil
 }
 
-func (c *Client) CreateColonyK3sContainer(ctx context.Context, loadBalancerIP, loadBalancerInterface, pwd string) error {
+func (c *Client) CreateColonyK3sContainer(ctx context.Context, colonyK3sBootstrapPath, colonyKubeconfigPath string) error {
 	log := logger.New(logger.Debug)
-
-	// TODO  tag a new repo for permanent housing, removes templates from database
-	colonyTemplateURL := "https://raw.githubusercontent.com/konstructio/colony/refs/heads/add-jobs/manifests/colony/colony.yaml.tmpl"
-	colonyTemplateYaml := filepath.Join(pwd, fmt.Sprintf("%s.tmpl", constants.ColonyYamlPath))
-
-	err := download.FileFromURL(colonyTemplateURL, colonyTemplateYaml)
-	if err != nil {
-		return fmt.Errorf("error downloading file: %w", err)
-	}
-
-	log.Infof("downloaded colony.yaml successfully to %q", colonyTemplateYaml)
-
-	err = hydrateTemplate(pwd, ColonyTokens{
-		LoadBalancerIP:        loadBalancerIP,
-		LoadBalancerInterface: loadBalancerInterface,
-	})
-	if err != nil {
-		return fmt.Errorf("error hydrating template: %w", err)
-	}
 
 	// check for an existing colony-k3s container
 	k3sColonyContainer, err := c.getColonyK3sContainer(ctx)
@@ -144,24 +117,24 @@ func (c *Client) CreateColonyK3sContainer(ctx context.Context, loadBalancerIP, l
 	io.Copy(os.Stdout, reader)
 
 	env := []string{
-		fmt.Sprintf("K3S_KUBECONFIG_OUTPUT=%s", constants.KubeconfigDockerPath),
+		fmt.Sprintf("K3S_KUBECONFIG_OUTPUT=%s", colonyKubeconfigPath),
 		"K3S_KUBECONFIG_MODE=666",
 	}
 
 	mounts := []mount.Mount{
 		{
 			Type:   mount.TypeBind,
-			Source: pwd,
+			Source: colonyKubeconfigPath,
 			Target: "/output",
 		},
-		{
-			Type:   mount.TypeVolume,
-			Source: "k3s-server",
-			Target: "/var/lib/rancher/k3s",
-		},
+		// {
+		// 	Type:   mount.TypeVolume,
+		// 	Source: "k3s-server",
+		// 	Target: "/var/lib/rancher/k3s",
+		// },
 		{
 			Type:   mount.TypeBind,
-			Source: filepath.Join(pwd, constants.ColonyYamlPath),
+			Source: colonyK3sBootstrapPath,
 			Target: fmt.Sprintf("/var/lib/rancher/k3s/server/manifests/%s", constants.ColonyYamlPath),
 		},
 		{
@@ -201,9 +174,9 @@ func (c *Client) CreateColonyK3sContainer(ctx context.Context, loadBalancerIP, l
 	waitInterval := 2 * time.Second
 	timeout := 15 * time.Second
 
-	log.Infof("Checking for file %s every %.2f seconds...", filepath.Join(pwd, constants.KubeconfigHostPath), waitInterval.Seconds())
+	log.Infof("Checking for file %s every %.0f seconds...", colonyKubeconfigPath, waitInterval.Seconds())
 
-	err = waitUntilFileExists(log, filepath.Join(pwd, constants.KubeconfigHostPath), waitInterval, timeout)
+	err = waitUntilFileExists(log, colonyKubeconfigPath, waitInterval, timeout)
 	if err != nil {
 		return fmt.Errorf("error waiting for kubeconfig file: %w", err)
 	}
@@ -236,27 +209,4 @@ func waitUntilFileExists(log *logger.Logger, filename string, interval, timeout 
 			return nil
 		}
 	}
-}
-
-func hydrateTemplate(pwd string, colonyTokens ColonyTokens) error {
-	tmpl, err := template.ParseFiles(filepath.Join(pwd, fmt.Sprintf("%s.tmpl", constants.ColonyYamlPath)))
-	if err != nil {
-		return fmt.Errorf("error parsing template: %w", err)
-	}
-
-	outputFile, err := os.Create(filepath.Join(pwd, constants.ColonyYamlPath))
-	if err != nil {
-		return fmt.Errorf("error creating output file: %w", err)
-	}
-	defer outputFile.Close()
-
-	err = tmpl.Execute(outputFile, &ColonyTokens{
-		LoadBalancerIP:        colonyTokens.LoadBalancerIP,
-		LoadBalancerInterface: colonyTokens.LoadBalancerInterface,
-	})
-	if err != nil {
-		return fmt.Errorf("error executing template: %w", err)
-	}
-
-	return nil
 }
